Large update of joel's code

+++ src/starvationevasion/io/CropCSVLoader.java
@@ -152,9 +152,9 @@ public class CropCSVLoader
   {
     for (EnumFood crop : EnumFood.values())
     {
-      int imports = agriculturalUnitTemp.getCropImport(crop);
-      int exports = agriculturalUnitTemp.getCropExport(crop);
-      int production = agriculturalUnitTemp.getCropProduction(crop);
+      long imports = agriculturalUnitTemp.getCropImport(crop);
+      long exports = agriculturalUnitTemp.getCropExport(crop);
+      long production = agriculturalUnitTemp.getCropProduction(crop);
       int land = agriculturalUnitTemp.getCropLand(crop);
       double yield = agriculturalUnitTemp.getCropYield(crop);
       double need = agriculturalUnitTemp.getCropNeedPerCapita(crop);
